Add history folder
