+++ app/commands/activate.js
@@ -4,7 +4,7 @@ import { SETTING_AI_ACTIVATED } from '../../constants/setting.js';
 import { t } from '../../languages/index.js';
 import storage from '../../storage/index.js';
 import Context from '../context.js';
-import { updateHistory } from '../histories.js';
+import { updateHistory } from '../history/index.js';
 
 /**
  * @param {Context} context
