+++ app/commands/deploy.js
@@ -3,7 +3,7 @@ import { COMMAND_DEPLOY } from '../../constants/command.js';
 import { t } from '../../languages/index.js';
 import { deploy } from '../../services/vercel.js';
 import Context from '../context.js';
-import { updateHistory } from '../histories.js';
+import { updateHistory } from '../history/index.js';
 
 /**
  * @param {Context} context
