+++ app/history/index.js
@@ -1,40 +1,4 @@
-// eslint-disable-next-line max-classes-per-file
-const MAX_RECORD_COUNT = 16;
-
-// FIXME
-class Record {
-  title;
-
-  text;
-
-  constructor({
-    title,
-    text,
-  }) {
-    this.title = title;
-    this.text = text;
-  }
-
-  toString() {
-    return `${this.title}: ${this.text}`;
-  }
-}
-
-class History {
-  records = [];
-
-  write(displayName, text) {
-    if (this.records.length >= MAX_RECORD_COUNT) {
-      this.records.shift();
-    }
-    this.records.push(new Record({ title: displayName, text }));
-    return this;
-  }
-
-  toString() {
-    return this.records.map((record) => record.toString()).join('\n');
-  }
-}
+import History from './history.js';
 
 const histories = new Map();
