+++ app/index.js
@@ -1,11 +1,9 @@
 import handleEvents from './handle-events.js';
 import { getPrompt, removePrompt, setPrompt } from './prompt/index.js';
-import { printHistories } from './histories.js';
 
 export {
   handleEvents,
   getPrompt,
   removePrompt,
   setPrompt,
-  printHistories,
 };
