+++ tests/default.test.js
@@ -2,7 +2,7 @@ import {
   afterEach, beforeEach, expect, test,
 } from '@jest/globals';
 import { getPrompt, handleEvents, removePrompt } from '../app/index.js';
-import { printHistories } from '../app/histories.js';
+import { printHistories } from '../app/history/index.js';
 import config from '../config/index.js';
 import {
   createEvents, TIMEOUT, MOCK_USER_01, MOCK_TEXT_OK,
